misc: surface_sam: Update based on module
Update Surface System Aggregator Module driver based on module.

Changes:
- Fix bug where second Surface Book 3 battery would not be set up
  correctly.
- Continued work on internal documentation and cleanup.

Links:
- SAM: https://github.com/linux-surface/surface-aggregator-module/commit/242667c594c691520196794fc05c198bad1245de

+++ drivers/misc/surface_sam/ssh_protocol.h
@@ -34,7 +34,8 @@
  * @buf: The pointer pointing to the data for which the CRC should be computed.
  * @len: The length of the data for which the CRC should be computed.
  *
- * Compute and return the CRC of the provided data, as used for SSH messages.
+ * Return: Returns the CRC computed on the provided data, as used for SSH
+ * messages.
  */
 static inline u16 ssh_crc(const u8 *buf, size_t len)
 {
@@ -45,9 +46,9 @@ static inline u16 ssh_crc(const u8 *buf, size_t len)
  * ssh_rqid_next_valid() - Return the next valid request ID.
  * @rqid: The current request ID.
  *
- * Compute and return the next valid request ID, following the current request
- * ID provided to this function. This function skips any request IDs reserved
- * for events.
+ * Return: Returns the next valid request ID, following the current request ID
+ * provided to this function. This function skips any request IDs reserved for
+ * events.
  */
 static inline u16 ssh_rqid_next_valid(u16 rqid)
 {
@@ -76,7 +77,7 @@ static inline bool ssh_rqid_is_event(u16 rqid)
  * ssh_tc_to_rqid() - Convert target category to its corresponding request ID.
  * @tc: The target category to convert.
  */
-static inline int ssh_tc_to_rqid(u8 tc)
+static inline u16 ssh_tc_to_rqid(u8 tc)
 {
 	return tc;
 }
